+++ changelog.md
@@ -5,6 +5,9 @@
 - Execute as much in parallel as possible
 - Make CLI output pretty
 - Add branch to nuxtus git repo for easier debugging
+- Prompt for database type
+- Automatically create SQLite database
+- Automatically start on successful install (if using SQLite)
 
 # 1.0.1