+++ changelog.md
@@ -1,5 +1,14 @@
 # Create Nuxtus
 
+# 1.0.2
+
+- Execute as much in parallel as possible
+- Make CLI output pretty
+- Add branch to nuxtus git repo for easier debugging
+- Prompt for database type
+- Automatically create SQLite database
+- Automatically start on successful install (if using SQLite)
+
 # 1.0.1
 
 - Put .env in client and server .gitignore files